第十七次提交 完成博客文章模块开发,集合开源的 md-editor-v3 编辑器,实现在线md文档编写与预览

+++ weblog-springboot/weblog-module-admin/src/main/java/com/qiccc/weblog/admin/service/impl/AdminCategoryServiceImpl.java
@@ -6,7 +6,9 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.qiccc.weblog.admin.model.vo.category.AddCategoryReqVO;
 import com.qiccc.weblog.admin.model.vo.category.DeleteCategoryReqVO;
 import com.qiccc.weblog.admin.service.AdminCategoryService;
+import com.qiccc.weblog.common.domain.dos.ArticleCategoryRelDO;
 import com.qiccc.weblog.common.domain.dos.CategoryDO;
+import com.qiccc.weblog.common.domain.mapper.ArticleCategoryRelMapper;
 import com.qiccc.weblog.common.domain.mapper.CategoryMapper;
 import com.qiccc.weblog.common.enums.ResponseCodeEnum;
 import com.qiccc.weblog.common.exception.BizException;
@@ -31,7 +33,8 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
-
+    @Autowired
+    private ArticleCategoryRelMapper articleCategoryRelMapper;
 
     /**
      * 添加分类
@@ -98,6 +101,14 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
     public Response deleteCategory(DeleteCategoryReqVO deleteCategoryReqVO) {
         // 分类 ID
         Long categoryId = deleteCategoryReqVO.getId();
+
+        // 校验该分类下是否已经有文章，若有，则提示需要先删除分类下所有文章，才能删除
+        ArticleCategoryRelDO articleCategoryRelDO = articleCategoryRelMapper.selectOneByCategoryId(categoryId);
+
+        if (Objects.nonNull(articleCategoryRelDO)) {
+            log.warn("==> 此分类下包含文章，无法删除，categoryId: {}", categoryId);
+            throw new BizException(ResponseCodeEnum.CATEGORY_CAN_NOT_DELETE);
+        }
 
         // 删除分类
         categoryMapper.deleteById(categoryId);
